Polishing: created generator utils, docstring, signatures
