+++ mlcolvar/core/loss/__init__.py
@@ -16,7 +16,6 @@ __all__ = [
     "SmartDerivatives",
     "compute_descriptors_derivatives",
     "GeneratorLoss",
-    "compute_eigenfunctions",
 ]
 
 from .mse import MSELoss, mse_loss
@@ -26,4 +25,4 @@ from .elbo import ELBOGaussiansLoss, elbo_gaussians_loss
 from .autocorrelation import AutocorrelationLoss, autocorrelation_loss
 from .fisher import FisherDiscriminantLoss, fisher_discriminant_loss
 from .committor_loss import CommittorLoss, committor_loss, SmartDerivatives, compute_descriptors_derivatives
-from .generator_loss import GeneratorLoss, compute_eigenfunctions
+from .generator_loss import GeneratorLoss
