+++ mlcolvar/cvs/generator/__init__.py
@@ -1,3 +1,4 @@
-__all__ = ["Generator"]
+__all__ = ["Generator", "compute_eigenfunctions", "forecast_state_occupation"]
 
 from .generator import *
+from .utils import *
