+++ mlcolvar/tests/test_cvs_generator.py
@@ -1,5 +1,5 @@
 from mlcolvar.cvs.generator.generator import test_generator
-from mlcolvar.core.loss.generator_loss import test_forecast_state_occupation
+from mlcolvar.cvs.generator.utils import test_forecast_state_occupation
 
 if __name__ == "__main__":
     test_generator()
